up 29-2 - documentado routineController y control errores, solucionado botones suscribir y desuscribir

+++ routes/web.php
@@ -31,31 +31,31 @@ Route::get('/', function () {
 Route::middleware('auth', 'verified')->group(function () {
     Route::get('/rutinas', [RoutineController::class, 'index'])->name('routine.index');
     Route::get('/rutina/{id}', [RoutineController::class, 'show'])->name('routine.show');
-    /*Route::get('/prueba/{id}',[UserController::class,'getUserRoutine'])->name('index');  */
-    Route::delete('/rutinas/eliminar/{id}', [RoutineController::class, 'destroy'])->name('routines.destroy')->middleware('role:trainer');
+
+    Route::delete('/rutinas/eliminar/{id}', [RoutineController::class, 'destroy'])->name('routines.destroy')->middleware('role:trainer|admin');
     // Ruta para el metodo que muestra la vista para crear una rutina nueva, solo pueden acceder usuarios autenticados con rol trainer
     Route::get('/rutinas/crear-rutina', [RoutineController::class, 'create'])->middleware('role:trainer')->name('rutinas.create');
     // Ruta para guardar una nueva rutina
-    Route::post('/rutinas', [RoutineController::class, 'store'])->middleware('role:trainer')->name('rutinas.store');
+    Route::post('/rutinas', [RoutineController::class, 'store'])->middleware('role:trainer|admin')->name('rutinas.store');
     // ruta para update de rutinas
-    Route::get('/rutinas/editar/{id}', [RoutineController::class, 'edit'])->name('rutinas.edit');
+    Route::get('/rutinas/editar/{id}', [RoutineController::class, 'edit'])->name('rutinas.edit')->middleware('role:trainer|admin');
     Route::put('/rutinas/editar/{id}', [RoutineController::class, 'update'])->middleware('role:trainer|admin')->name('rutinas.update');
 });
 
 // Rutas usuario
 Route::middleware(['auth', 'verified'])->group(function () {
-    Route::get('/entrenador/{id}/rutinas', [UserController::class, 'obtainCreatedRoutines'])->name('users.trainerRoutines');
-    Route::post('/suscribir-usuario/{user_id}/rutina/{routine_id}', [UserController::class, 'subscribeUserToRoutine']);
-    Route::post('/des-suscribir-usuario/{user_id}/rutina/{routine_id}', [UserController::class, 'unSubscribeUserFromRoutine']);
+    Route::get('/entrenador/{id}/rutinas', [UserController::class, 'obtainCreatedRoutines'])->name('users.trainerRoutines')->middleware('role:trainer');
+    Route::post('/suscribir-usuario/{user_id}/rutina/{routine_id}', [UserController::class, 'subscribeUserToRoutine'])->middleware('role:user');
+    Route::post('/des-suscribir-usuario/{user_id}/rutina/{routine_id}', [UserController::class, 'unSubscribeUserFromRoutine'])->middleware('role:user');
 
     // Middleware de role admin  (solo pueden entrar administradores)
     Route::get('administrar-usuarios/', [UserController::class, 'index'])
         ->name('users.manageUsers')
-        ->middleware('role:admin'); // Asegúrate de que el rol sea exactamente 'admin'
+        ->middleware('role:admin');
     // Ruta para eliminar usuarios
     Route::delete('/administrar-usuarios/eliminar/{id}', [UserController::class, 'destroy'])->name('admin.users.destroy')->middleware('role:admin');
     // Administracion de clientes y sus entrenamientos
-    Route::get('administrar-clientes/', [UserController::class, 'getUsersWithTheirRoutines'])->name('users.trainerClients')->middleware(['role:trainer', 'auth', 'verified']);
+    Route::get('administrar-clientes/', [UserController::class, 'getUsersWithTheirRoutines'])->name('users.trainerClients')->middleware('role:trainer');
     Route::get('administrar-clientes/ver-entrenamientos/{userId}', [UserController::class, 'listUserWorkouts'])->name('users.listUserTrainings')->middleware('role:trainer');
 });
 // SUBSCRIPCIONES
